# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Readiness Analyzer API for Shopify stores
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/{storeId}/analyze:
    post:
      operationId: analyzeStore
      tags: [analysis]
      summary: Trigger full store analysis
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJob"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{storeId}/summary:
    get:
      operationId: getStoreSummary
      tags: [analysis]
      summary: Get store AI readiness summary and scores
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSummary"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /stores/{storeId}/gaps:
    get:
      operationId: listGaps
      tags: [analysis]
      summary: Get all detected gaps ranked by severity
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of gaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gap"
  /stores/{storeId}/consistency:
    get:
      operationId: getConsistencyReport
      tags: [analysis]
      summary: Get store-wide consistency report
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Consistency report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsistencyReport"
  /stores/{storeId}/benchmark:
    get:
      operationId: getBenchmark
      tags: [analysis]
      summary: Compare store against ideal AI-ready benchmark
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Benchmark comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkReport"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AnalysisJob:
      type: object
      properties:
        jobId:
          type: string
        storeId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        startedAt:
          type: string
        message:
          type: string
      required:
        - jobId
        - storeId
        - status
        - startedAt
        - message
    StoreSummary:
      type: object
      properties:
        storeId:
          type: string
        storeName:
          type: string
        overallScore:
          type: number
        clarityScore:
          type: number
        completenessScore:
          type: number
        trustScore:
          type: number
        tagScore:
          type: number
        consistencyScore:
          type: number
        policyScore:
          type: number
        totalProducts:
          type: integer
        analyzedProducts:
          type: integer
        criticalIssues:
          type: integer
        mediumIssues:
          type: integer
        lowIssues:
          type: integer
        pendingFixes:
          type: integer
        appliedFixes:
          type: integer
        lastAnalyzed:
          type: ["string", "null"]
      required:
        - storeId
        - storeName
        - overallScore
        - clarityScore
        - completenessScore
        - trustScore
        - tagScore
        - consistencyScore
        - policyScore
        - totalProducts
        - analyzedProducts
        - criticalIssues
        - mediumIssues
        - lowIssues
        - pendingFixes
        - appliedFixes
...
    Gap:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        productId:
          type: ["string", "null"]
        productTitle:
          type: ["string", "null"]
        category:
          type: string
          enum: [clarity, completeness, trust, tags, policy, consistency]
        severity:
          type: string
          enum: [high, medium, low]
        title:
          type: string
        description:
          type: string
        suggestion:
          type: string
        isFixed:
          type: boolean
      required:
        - id
        - storeId
        - category
        - severity
        - title
        - description
        - suggestion
        - isFixed
    ConsistencyIssue:
      type: object
      properties:
        type:
          type: string
          enum: [tone, structure, formatting, missing_section]
        description:
          type: string
        affectedProductCount:
          type: integer
        examples:
          type: array
          items:
            type: string
      required:
        - type
        - description
        - affectedProductCount
        - examples
    ConsistencyReport:
      type: object
      properties:
        storeId:
          type: string
        overallConsistencyScore:
          type: number
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyIssue"
        suggestedStructure:
          type: array
          items:
            type: string
        analyzedAt:
          type: ["string", "null"]
      required:
        - storeId
        - overallConsistencyScore
        - issues
        - suggestedStructure
    BenchmarkDimension:
      type: object
      properties:
        name:
          type: string
        storeScore:
          type: number
        benchmarkScore:
          type: number
        gap:
          type: number
        priority:
          type: string
          enum: [high, medium, low]
      required:
        - name
        - storeScore
        - benchmarkScore
        - gap
        - priority
    BenchmarkReport:
      type: object
      properties:
        storeId:
          type: string
        overallStoreScore:
          type: number
        overallBenchmarkScore:
          type: number
        overallGap:
          type: number
        dimensions:
          type: array
          items:
            $ref: "#/components/schemas/BenchmarkDimension"
        topImprovements:
          type: array
          items:
            type: string
      required:
        - storeId
        - overallStoreScore
        - overallBenchmarkScore
        - overallGap
        - dimensions
        - topImprovements
...